# client/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenCap Stack API",
    "description": "Cap table and equity management REST API aligned with OCTA v2.0 schema. Manage stakeholders, share classes, equity grants, SAFE notes, 409A valuations, and investor reports.",
    "version": "1.0.0",
    "contact": {
      "name": "OpenCap Stack",
      "url": "https://opencapstack.com/support",
      "email": "support@opencapstack.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/Open-Cap-Stack/opencapstack/blob/main/LICENSE"
    },
    "x-logo": "https://opencapstack.com/logo.png"
  },
  "servers": [
    { "url": "https://api.opencapstack.com/api/v1", "description": "Production" },
    { "url": "http://localhost:3000/api/v1", "description": "Local development" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "status": { "type": "integer" },
              "docs": { "type": "string", "format": "uri" }
            }
          }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login and obtain JWT",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns { token, refreshToken }" },
          "401": { "$ref": "#/components/schemas/Error" }
        }
      }
    },
...
  }
}